refactor(rate-of-closure): split simulation tab controller
